restructure solar/lunar eclipse
